+++ src/components/Pages/Profile/Edit/ProfileEditButton/ProfileEditButton.tsx
@@ -1,12 +1,11 @@
+import { useEffect } from "react";
+import { useRouter } from "next/navigation";
+
 import { Button } from "@/components";
 import { ButtonStyle } from "@/components/Button/Button.types";
 import { PROFILE_EDIT_TEXT } from "@/constants/messages";
-import { useRouter } from "next/navigation";
-import { useProfileEditMutation } from "@/hooks/api/profileEdit/useProfileEditMutation";
-import useTokenStore from "@/store/user/useTokenStore";
-import useProfileEditStore from "@/store/user/useProfileEditStore";
-import { useEffect } from "react";
-import useUserInfoStore from "@/store/user/useUserInfoStore";
+import useProfileEditMutation from "@/hooks/api/profileEdit/useProfileEditMutation";
+import { useTokenStore, useProfileEditStore, useUserInfoStore } from "@/store";
 import * as S from "./ProfileEditButton.style";
 
 export default function ProfileEditButton() {
