+++ src/components/Pages/Signup/SignupJob/SignupJob.tsx
@@ -1,7 +1,7 @@
 import { Chip } from "@/components";
 import { SIGNUP_TEXT, STAR_ICON } from "@/constants/messages";
 import SignupJobs from "@/types/enum/signupJobs";
-import useSignupUserStore from "@/store/user/useSignupUserStore";
+import { useSignupUserStore } from "@/store";
 import * as S from "./SignupJob.style";
 
 export default function SignupJob() {
