+++ src/components/Toolbar/Toolbar.tsx
@@ -1,11 +1,14 @@
 import { useState } from "react";
+
 import { Button, ButtonList, ChipList, ContextMenu, Tab } from "@/components";
 import resetIcon from "@/assets/icons/reset-left-line.svg";
 import arrowDown from "@/assets/icons/arrow-down.svg";
 import checkLineIcon from "@/assets/icons/check-line.svg";
-import useChipStore from "@/store/component/useChipStore";
-import useContextMenuStore from "@/store/common/useContextMenuStore";
-import useDesignSystemFilterStore from "@/store/designSystem/useDesignSystemFilterStore";
+import {
+  useChipStore,
+  useContextMenuStore,
+  useDesignSystemFilterStore,
+} from "@/store";
 import * as S from "./Toolbar.style";
 import { ButtonStyle } from "../Button/Button.types";
